ANDROID: KVM: arm64: Convert kvm_pinned_pages to an interval-tree
Here we are, the last straw... the maple-tree is absolutely terrible for
what we need and it is time to get rid of it. With the upcoming set of
fixes for memory relinquish with huge-mappings, we need to be able to
split a pinned_page under the mmu write_lock. This is just too
complicated with the maple-tree while it is a piece of cake with an
interval tree.

Bug: 421013437
Bug: 419548963
Change-Id: I981b5d875085e1d2f7b4ebe2560c8b6ea3cbae88
Signed-off-by: Vincent Donnefort <vdonnefort@google.com>
(cherry picked from commit e56d181356a44521a96fb844cfc057ad8030dc8d)
Signed-off-by: Will McVicker <willmcvicker@google.com>
[willmcvicker: fix merge conflict in abi_gki_aarch64.stg.allowed_breaks]