zram: Set default compressor to LZ4 and initialize default disksize to 16GB
Change the default zram compressor from LZ4KD to LZ4 in the Kconfig to
align with common use cases and improve compatibility. Additionally,
introduce a default disksize of 16GB in zram_add() to simplify initial
device setup. The disksize is set during initialization, and the
compression backends are created accordingly. Update the log message to
reflect the default size. Also, fix a minor typo in a comment for clarity.

Signed-off-by: brokestar233 <3765589194@qq.com>