mm: zswap: Fix zswap shrinker data read issue on cgroup v1
When memory cgroup is enabled but running on cgroup v1, the zswap
shrinker fails to read data due to incorrect memcg statistics handling.
This is because the original code does not account for cgroup v1 when
checking memcg stats.

Add a condition to check if the memory cgroup subsystem is on the default
hierarchy (cgroup v2) before accessing memcg stats, ensuring correct
behavior for both cgroup v1 and v2.

Signed-off-by: brokestar233 <3765589194@qq.com>