power: Revert "Add wakeup event handling during filesystem sync in suspend"
This reverts commit fdc5abc0fed051c6c47c2f5844565a761f322b19.

The patch introduced functionality to handle wakeup events during the filesystem
sync phase of suspend, allowing aborts to improve responsiveness. However, it has
been found to cause a system deadlock on memory, resulting in a QUALCOMM CrashDump
Mode panic in the android.bg process. The call stack indicates an out-of-memory
condition triggered during page allocation:

  panic+0x178/0x38c
  out_of_memory+0x2a0/0x340
  __alloc_pages_may_oom+0xd0/0x2c0
  __alloc_pages+0x758/0x1364
  filemap_fault+0x3b4/0xd28

To restore system stability, this patch is reverted until the root cause of the
memory deadlock is identified and addressed.

Link: https://lore.kernel.org/all/20250821004237.2712312-1-wusamuel@google.com/
Bug: 315272811
Signed-off-by: brokestar233 <3765589194@qq.com>